Transferred the read method to the new class
